Move variable roles to separate file
