+++ blocks/graph.py
@@ -10,6 +10,7 @@ from theano.gof.sched import make_dependence_cmp, sort_apply_nodes
 from theano.sandbox.rng_mrg import MRG_RandomStreams
 
 from blocks import config
+from blocks.roles import add_role, AUXILIARY
 from blocks.utils import (is_graph_input, is_shared_variable, dict_union,
                           shared_like)
 
@@ -157,39 +158,6 @@ class ComputationGraph(object):
                                                          value_holders)])
 
 
-def add_role(var, role):
-    r"""Add a role to a given Theano variable.
-
-    Parameters
-    ----------
-    var : Theano variable
-        The variable to assign the new role to.
-    role : :class:`VariableRole` instance
-
-    Notes
-    -----
-    Some roles are subroles of others (e.g. :const:`WEIGHTS` is a subrole
-    of :const:`PARAMETER`). This function will not add a role if a more
-    specific role has already been added. If you need to replace a role
-    with a parent role (e.g. replace :const:`WEIGHTS` with
-    :const:`PARAMETER`) you must do so manually.
-
-    Examples
-    --------
-    >>> from theano import tensor
-    >>> W = tensor.matrix()
-    >>> from blocks.bricks import WEIGHTS
-    >>> add_role(W, WEIGHTS)
-    >>> print(*W.tag.roles)
-    WEIGHTS
-
-    """
-    roles = getattr(var.tag, 'roles', [])
-    roles = [old_role for old_role in roles
-             if not isinstance(role, old_role.__class__)] + [role]
-    var.tag.roles = roles
-
-
 def add_annotation(var, annotation):
     annotations = getattr(var.tag, 'annotations', [])
     if any(old_annotation.__class__ == annotation.__class__
@@ -265,6 +233,7 @@ class Annotation(object):
         Examples
         --------
         >>> from blocks.bricks.base import application, Brick
+        >>> from blocks.roles import COST
         >>> from blocks.utils import shared_floatx_zeros
         >>> class Foo(Brick):
         ...     def _allocate(self):
@@ -325,38 +294,3 @@ def apply_noise(graph, variables, level, seed=None):
         replace[variable] = (variable +
                              rng.normal(variable.shape, std=level))
     return graph.replace(replace)
-
-
-class VariableRole(object):
-    def __str__(self):
-        return self.__class__.__name__[:-4].upper()
-
-
-class InputRole(VariableRole):
-    pass
-
-INPUT = InputRole()
-
-
-class OutputRole(VariableRole):
-    pass
-
-OUTPUT = OutputRole
-
-
-class CostRole(VariableRole):
-    pass
-
-COST = CostRole()
-
-
-class ParameterRole(VariableRole):
-    pass
-
-PARAMETER = ParameterRole()
-
-
-class AuxiliaryRole(VariableRole):
-    pass
-
-AUXILIARY = AuxiliaryRole()
