+++ blocks/roles.py
@@ -1,0 +1,78 @@
+def add_role(var, role):
+    r"""Add a role to a given Theano variable.
+
+    Parameters
+    ----------
+    var : Theano variable
+        The variable to assign the new role to.
+    role : :class:`VariableRole` instance
+
+    Notes
+    -----
+    Some roles are subroles of others (e.g. :const:`WEIGHTS` is a subrole
+    of :const:`PARAMETER`). This function will not add a role if a more
+    specific role has already been added. If you need to replace a role
+    with a parent role (e.g. replace :const:`WEIGHTS` with
+    :const:`PARAMETER`) you must do so manually.
+
+    Examples
+    --------
+    >>> from theano import tensor
+    >>> W = tensor.matrix()
+    >>> from blocks.bricks import WEIGHTS
+    >>> add_role(W, WEIGHTS)
+    >>> print(*W.tag.roles)
+    WEIGHTS
+
+    """
+    roles = getattr(var.tag, 'roles', [])
+    roles = [old_role for old_role in roles
+             if not isinstance(role, old_role.__class__)] + [role]
+    var.tag.roles = roles
+
+
+class VariableRole(object):
+    def __str__(self):
+        return self.__class__.__name__[:-4].upper()
+
+
+class InputRole(VariableRole):
+    pass
+
+INPUT = InputRole()
+
+
+class OutputRole(VariableRole):
+    pass
+
+OUTPUT = OutputRole
+
+
+class CostRole(VariableRole):
+    pass
+
+COST = CostRole()
+
+
+class ParameterRole(VariableRole):
+    pass
+
+PARAMETER = ParameterRole()
+
+
+class AuxiliaryRole(VariableRole):
+    pass
+
+AUXILIARY = AuxiliaryRole()
+
+
+class WeightsRole(ParameterRole):
+    pass
+
+WEIGHTS = WeightsRole()
+
+
+class BiasesRole(ParameterRole):
+    pass
+
+BIASES = BiasesRole()
